Bug 1229593 - Move custom menu items to MenuHelper
