+++ Client/Application/AppDelegate.swift
@@ -57,6 +57,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Start the keyboard helper to monitor and cache keyboard state.
         DynamicFontHelper.defaultHelper.startObserving()
 
+        log.debug("Setting custom menu items…")
+        MenuHelper.defaultHelper.setItems()
+
         log.debug("Creating Sync log file…")
         let logDate = NSDate()
         // Create a new sync log file on cold app launch. Note that this doesn't roll old logs.
